Basic graphics functionality; code cleanup
