+++ graphics/packed_graphics.c
@@ -1,0 +1,26 @@
+/*
+ * Packed Graphics Interface
+ * 
+ * @Company
+ *   1o1 Oscilloscope Team
+ * 
+ * @File Name
+ *   packed_graphics.c
+ * 
+ * @Summary
+ *   Contains predefined bitmaps in packed_graphics_t format.
+*/
+
+#include "packed_graphics.h"
+
+#include <stdint.h>
+
+#include "../drivers/ssd1305.h"
+
+
+uint8_t logo_bytes[] = {0,0,0,0,0,0,0,0,0,0,0,0,0,0,128,3,0,0,0,0,0,0,128,7,0,0,0,0,0,0,128,15,0,0,0,0,0,0,0,31,254,255,255,255,255,255,255,63,254,255,255,255,255,255,255,127,254,255,255,255,255,255,255,127,254,255,255,255,255,255,255,63,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,240,255,0,0,0,0,30,0,254,255,7,0,0,0,30,128,255,255,31,0,0,0,30,224,255,255,127,0,0,0,30,248,15,0,255,1,0,0,30,252,3,0,252,3,0,0,30,254,0,0,240,7,0,0,30,63,0,0,192,15,0,0,158,31,0,0,128,31,0,0,222,15,0,0,0,63,0,0,206,7,0,0,0,126,0,0,238,227,0,0,0,124,0,0,254,225,1,0,0,248,0,0,254,192,15,0,0,240,0,0,254,128,127,0,0,240,1,0,126,0,254,3,0,224,1,0,126,0,240,31,0,224,3,0,62,0,128,255,0,192,3,0,62,0,0,252,7,192,3,0,62,0,0,224,31,128,7,0,30,0,0,0,127,128,7,0,30,0,0,0,248,128,7,0,30,0,0,0,224,128,7,0,30,0,0,0,192,129,7,0,30,0,0,0,224,128,7,0,30,0,0,0,248,0,7,0,30,0,0,0,63,0,7,0,30,0,0,224,31,0,7,0,30,0,0,252,3,128,7,0,30,0,128,255,0,128,7,0,30,0,240,15,0,128,7,0,30,0,254,1,0,128,7,0,30,128,63,0,0,128,7,0,62,192,7,0,0,192,7,0,62,224,1,0,0,192,3,0,62,112,0,0,0,192,3,0,126,112,0,0,0,224,3,0,126,224,1,0,0,224,1,0,254,224,7,0,0,240,1,0,254,129,63,0,0,248,0,0,238,1,254,1,0,248,0,0,254,3,248,15,0,124,0,0,222,7,192,127,0,62,0,0,158,15,0,254,1,31,0,0,30,31,0,240,131,31,0,0,30,127,0,0,225,15,0,0,30,252,0,0,240,7,0,0,30,248,7,0,254,1,0,0,30,240,63,192,255,0,0,0,30,192,255,255,127,0,0,0,30,128,255,255,31,0,0,0,30,0,252,255,3,0,0,0,30,0,192,63,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,0,0,30,0,0,0,0,0,64,0,30,224,255,255,255,255,127,0,254,255,255,255,255,255,255,15,252,255,255,255,255,255,255,15,248,255,255,255,255,255,255,127,240,255,255,255,255,255,255,15,0,248,255,255,255,255,127,0,0,0,0,0,0,0,96,0,0,0,0,0,0,0,64,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0};
+packed_graphics_t graphics_logo = {
+	96,
+	8,
+	logo_bytes
+};
