+++ scope_display.c
@@ -1,0 +1,14 @@
+/*
+ * Graphical Oscilloscope Display
+ * 
+ * @Company
+ *   1o1 Oscilloscope Team
+ * 
+ * @File Name
+ *   scope_display.c
+ * 
+ * @Summary
+ *   Contains the glue logic to run the display for the oscilloscope.
+*/
+
+#include "scope_display.h"
