+++ scope_display.h
@@ -1,0 +1,31 @@
+/*
+ * Graphical Oscilloscope Display
+ * 
+ * @Company
+ *   1o1 Oscilloscope Team
+ * 
+ * @File Name
+ *   scope_display.h
+ * 
+ * @Summary
+ *   Contains the glue logic to run the display for the oscilloscope.
+*/
+
+#ifndef SCOPE_DISPLAY_H
+#define	SCOPE_DISPLAY_H
+
+#ifdef	__cplusplus
+extern "C"
+{
+#endif
+
+
+
+
+
+#ifdef	__cplusplus
+}
+#endif
+
+#endif	/* SCOPE_DISPLAY_H */
+
